Added styling for the fetched data
Removed unnecessary code and added styling to the fetched data

+++ shared/src/lib.rs
@@ -32,7 +32,7 @@ use serde::{Serialize, Deserialize};
         "cod":200
      */
 
-#[derive(Deserialize, Serialize, Debug, PartialEq)]
+#[derive(Clone, Deserialize, Serialize, Debug, PartialEq)]
 pub struct WeatherData{
     pub id: i32,  
     pub name: String,
@@ -43,7 +43,7 @@ pub struct WeatherData{
     pub cod: i32
 }
 
-#[derive(Deserialize, Serialize, Debug, Default, PartialEq)]
+#[derive(Clone, Copy, Deserialize, Serialize, Debug, Default, PartialEq)]
 pub struct Main{
     pub humidity: i32,
     pub temp: f32,
@@ -52,16 +52,16 @@ pub struct Main{
     pub temp_max: f32,
 }
 
-#[derive(Deserialize, Serialize, Debug, Default, PartialEq)]
+#[derive(Clone, Copy, Deserialize, Serialize, Debug, Default, PartialEq)]
 pub struct Sys{
     pub sunrise: i64,
     pub sunset: i64,
 }
 
-#[derive(Deserialize, PartialEq, Serialize, Debug, Default)]
+#[derive(Clone, Copy, Deserialize, PartialEq, Serialize, Debug, Default)]
 pub struct Coords{
-    pub lon: f32,
-    pub lat: f32
+    pub lon: f64,
+    pub lat: f64
 }
 
 impl Default for WeatherData{
